+++ cucumber-junit-platform-engine/src/main/java/io/cucumber/junit/platform/engine/DefaultNamingStrategyProvider.java
@@ -11,6 +11,7 @@ import java.util.function.Supplier;
 
 import static io.cucumber.junit.platform.engine.Constants.JUNIT_PLATFORM_LONG_NAMING_STRATEGY_EXAMPLE_NAME_PROPERTY_NAME;
 import static io.cucumber.junit.platform.engine.Constants.JUNIT_PLATFORM_SHORT_NAMING_STRATEGY_EXAMPLE_NAME_PROPERTY_NAME;
+import static io.cucumber.junit.platform.engine.Constants.JUNIT_PLATFORM_SUREFIRE_NAMING_STRATEGY_EXAMPLE_NAME_PROPERTY_NAME;
 
 enum DefaultNamingStrategyProvider {
     LONG {
@@ -30,6 +31,16 @@ enum DefaultNamingStrategyProvider {
                     .map(DefaultNamingStrategyProvider::parseStrategy)
                     .orElse(DefaultNamingStrategyProvider::exampleNumberAndPickleIfParameterizedStrategy)
                     .apply(DefaultNamingStrategyProvider::shortStrategy);
+        }
+    },
+
+    SUREFIRE {
+        @Override
+        NamingStrategy create(ConfigurationParameters configuration) {
+            return configuration.get(JUNIT_PLATFORM_SUREFIRE_NAMING_STRATEGY_EXAMPLE_NAME_PROPERTY_NAME)
+                    .map(DefaultNamingStrategyProvider::parseStrategy)
+                    .orElse(DefaultNamingStrategyProvider::exampleNumberAndPickleIfParameterizedStrategy)
+                    .apply(DefaultNamingStrategyProvider::surefireStrategy);
         }
     };
 
@@ -124,5 +135,44 @@ enum DefaultNamingStrategyProvider {
         }
 
         return builder.toString();
+    }
+
+    private static String surefireStrategy(Node node, String currentNodeName) {
+        // Surefire uses the parents of test nodes to determine the class name.
+        // As we want the class name to match the feature name we name the
+        // parents of the test containing nodes after the feature.
+        if (node instanceof Node.Examples || node instanceof Node.Rule) {
+            return nameOrKeyword(findFeature(node));
+        }
+        // Scenarios and examples names are used by surefire to populate the
+        // testname. We want this to be long, but without the feature name
+        // because that will be used for the class name
+        if (node instanceof Node.Scenario || node instanceof Node.Example) {
+            return longStrategyWithoutFeatureName(node, currentNodeName);
+        }
+        // Everything else, can be short, will be ignored by surefire.
+        return shortStrategy(node, currentNodeName);
+    }
+
+    private static String longStrategyWithoutFeatureName(Node node, String currentNodeName) {
+        StringBuilder builder = new StringBuilder();
+        builder.append(currentNodeName);
+        node = node.getParent().orElse(null);
+
+        while (node != null && !(node instanceof Node.Feature)) {
+            builder.insert(0, " - ");
+            builder.insert(0, nameOrKeyword(node));
+            node = node.getParent().orElse(null);
+        }
+        return builder.toString();
+    }
+
+    private static Node findFeature(Node node) {
+        Node candidate = node.getParent().orElse(null);
+        while (candidate != null) {
+            node = candidate;
+            candidate = node.getParent().orElse(null);
+        }
+        return node;
     }
 }